Added Process Next/Previous functionality. Removed 4th DataViewGrid with "Matched by Distance with Threshold applied" in favor of a "Matched by Distance" grid with a checkbox that applies the threshold. Initial version of color matching algorithm used and "Matched" table populated. Indexed Threshold Percentage is now double (from int).

+++ trunk/ImagePalette/ImagePalette/ImagePaletteParameters.cs
@@ -84,11 +84,11 @@ namespace ImagePalette
             set { SetProperty<int>(ref distance, value); }
         }
 
-        private int thresholdIndexed;
-        public int ThresholdIndexed
+        private double thresholdIndexed;
+        public double ThresholdIndexed
         {
             get { return thresholdIndexed; }
-            set { SetProperty<int>(ref thresholdIndexed, value); }
+            set { SetProperty<double>(ref thresholdIndexed, value); }
         }
 
         private int thresholdMatched;
